% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dataZvE.R
\name{clean_dataZvE}
\alias{clean_dataZvE}
\title{Daten aus Lohn und EkStbericht extrahieren}
\usage{
clean_dataZvE(data_eink, nominallohn_fkt = 1)
}
\arguments{
\item{data_eink}{Tabelle 73111-02 aus Lohn- und Einkommenssteuerbericht hat Variablen "Zu_versteuerndes_Einkommen", "Grund-Splittingtabelle", "Steuerpflichtige_Anzahl", "Betrag_EUR". Kann weitere haben}

\item{nominallohn_fkt}{Faktor für Verschiebung der Nominallöhne e.g. Nominallohn_fkt=(1+"Annahme Lohnwachstum 2025")\*"Nominallohnindex Mai 2024"/"Nominallohnindex 2019"}
}
\value{
dataframe
}
\description{
Daten aus Lohn und EkStbericht extrahieren
}
\examples{
\dontrun{
clean_dataZvE(data_eink,1)}
}
