% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Einkommenssteuer.R
\name{compute_Einkommensteuer}
\alias{compute_Einkommensteuer}
\title{Einkommenssteuer berechnen}
\usage{
compute_Einkommensteuer(
  zve,
  gfb = 11604,
  pg1s = 14,
  pg1e = 17005,
  pg2s = 23.97,
  pg2e = 66760,
  pp1s = 42,
  pp1e = 277826,
  pp2s = 45
)
}
\arguments{
\item{zve}{Zu versteuerndes Einkommen}

\item{gfb}{Grundfreibetrags}

\item{pg1s}{Startsatz erste Progressionszone}

\item{pg1e}{Ende erste Progressionszone}

\item{pg2s}{Startsatz zweite Progressionszone}

\item{pg2e}{Ende zweite Progressionszone}

\item{pp1s}{Satz erste Proportionalzone, Spitzensteuersatz}

\item{pp1e}{Ende erste Proportionalzone}

\item{pp2s}{Seite zweite Proportionalzone, Reichensteuersatz}
}
\value{
value
}
\description{
Einkommenssteuer berechnen
}
\examples{
compute_Einkommensteuer(30000)
}
