% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Gesamtkosten.R
\name{compute_Gesamtkosten}
\alias{compute_Gesamtkosten}
\title{Compute Gesamtkosten}
\usage{
compute_Gesamtkosten(data)
}
\arguments{
\item{data}{Dataframe Reform}
}
\value{
number
}
\description{
Compute Gesamtkosten
}
\examples{
\dontrun{
compute_Gesamtkosten(Reform2025_df)
}
}
