% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_wachstumseffekte.R
\name{compute_wachstumseffekte}
\alias{compute_wachstumseffekte}
\title{Wachstumseffekte}
\usage{
compute_wachstumseffekte(
  Kosten,
  impulse_response = c(0.25, 1.1, 2.35, 2.4),
  Jahre = 2025:2028,
  GDP = 4406,
  steuerelast = 0.97,
  steuerschätzung = c(982.4, 1024.9, 1063.2, 1133.8),
  inflationsan = 2
)
}
\arguments{
\item{GDP}{Aktuelles GDP}

\item{steuerelast}{Steuerelastizität (Default 0.97 basierend auf )}

\item{steuerschätzung}{Steuerschätzung Default von}

\item{inflationsan}{Angenommene Inflation für die nächsten Jahre (für Annahme, dass sich die Kosten entsprechend der Inflation entwickeln)}

\item{data}{Reformvorschlag}

\item{multiplikator}{Impulse Response für Tax Shock in Höhe 1% GDP}
}
\value{
dataframe
}
\description{
Wachstumseffekte
}
\examples{
compute_wachstumseffekte(10)
}
