% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataframeReform.R
\name{create_dataframeReform}
\alias{create_dataframeReform}
\title{Datensatz für Reformvorschlag}
\usage{
create_dataframeReform(
  data_eink,
  gfb_reform = 11604,
  pg1s_reform = 14,
  pg1e_reform = 17005,
  pg2s_reform = 23.97,
  pg2e_reform = 66760,
  pp1s_reform = 42,
  pp1e_reform = 277826,
  pp2s_reform = 45,
  solib_base = 18130,
  solis_base = 5.5,
  solig_base = 11.9,
  gfb_base = 11604,
  pg1s_base = 14,
  pg1e_base = 17005,
  pg2s_base = 23.97,
  pg2e_base = 66760,
  pp1s_base = 42,
  pp1e_base = 277826,
  pp2s_base = 45,
  solib_reform = 18130,
  solis_reform = 5.5,
  solig_reform = 11.9
)
}
\arguments{
\item{data_eink}{Tabelle 73111-02 aus Lohn- und Einkommenssteuerbericht}

\item{gfb_reform}{Grundfreibetrag Reform}

\item{pg1s_reform}{Startsatz erste Progressionszone}

\item{pg1e_reform}{Ende erste Progressionszone}

\item{pg2s_reform}{Startsatz zweite Progressionszone}

\item{pg2e_reform}{Ende zweite Progressionszone}

\item{pp1s_reform}{Satz erste Proportionalzone, Spitzensteuersatz}

\item{pp1e_reform}{Ende erste Proportionalzone}

\item{pp2s_reform}{Satz zweite Proportionalzone, Reichensteuersatz}

\item{solib_base}{Solibeginn}

\item{solis_base}{Solisatz}

\item{solig_base}{Gleitfaktor Soli}

\item{gfb_base}{Grundfreibetrag}

\item{pg1s_base}{Startsatz erste Progressionszone}

\item{pg1e_base}{Ende erste Progressionszone}

\item{pg2s_base}{Startsatz zweite Progressionszone}

\item{pg2e_base}{Ende zweite Progressionszone}

\item{pp1s_base}{Satz erste Proportionalzone, Spitzensteuersatz}

\item{pp1e_base}{Ende erste Proportionalzone}

\item{pp2s_base}{Satz zweite Proportionalzone, Reichensteuersatz}

\item{solib_reform}{Solibeginn}

\item{solis_reform}{Solisatz}

\item{solig_reform}{Gleitfaktor Soli}
}
\value{
dataframe
}
\description{
Datensatz für Reformvorschlag
}
\examples{
\dontrun{
get_dataframeReform(data_eink)}
}
