% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dataframe_Entlastungen.R
\name{gen_dataframe_Entlastungen}
\alias{gen_dataframe_Entlastungen}
\title{Get Entlastungen}
\usage{
gen_dataframe_Entlastungen(data, n_groups = 10, formated = FALSE)
}
\arguments{
\item{data}{Dataframe Reform}

\item{n_groups}{Anzahl Gruppen}

\item{formated}{Soll ein rein numerische Datensatz oder ein bereits formatierter exportiert werden?}
}
\value{
dataframe or formated dataframe
}
\description{
Entlastungen (Kosten) insgesamt und nach n-til (im Status Quo der Steuerpflichtigen)
}
\examples{
\dontrun{
get_dataframe_Entlastungen(SPD_2025,10)
}
}
