% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_waterfallCost.R
\name{gen_dataframe_Waterfallplot}
\alias{gen_dataframe_Waterfallplot}
\title{Internal Function for WaterfallPlot}
\usage{
gen_dataframe_Waterfallplot(data_Wachstumseffekte, year = "2025", max_y = NULL)
}
\arguments{
\item{data_Wachstumseffekte}{dataframe Wachstumseffekte}

\item{year}{Jahr des Wachstumseffekts}

\item{max_y}{Obergrenze Y-Achse}
}
\value{
dataframe
}
\description{
Internal Function for WaterfallPlot
}
\examples{
\dontrun{gen_dataframe_Waterfallplot(data_Wachstumseffekte)}
}
