% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_plot_EntlastungEinkommensverteilung.R
\name{gen_plot_EntlastungEinkommensverteilung}
\alias{gen_plot_EntlastungEinkommensverteilung}
\title{Plot Entlastung Einkommensverteilung}
\usage{
gen_plot_EntlastungEinkommensverteilung(
  df,
  breaks = c(10000, 30000, 1e+05, 5e+05),
  max_x = 5e+05,
  width = 1,
  max_val = 8,
  interactive = F
)
}
\arguments{
\item{df}{dataframe Reform}

\item{breaks}{breaks X-Achse}

\item{width}{breite Gruppen}

\item{max_val}{Obergrenze Gruppe (symetrisch)}

\item{interactive}{Return interactive plot?}

\item{limit}{limit X-Achse}
}
\value{
ggplot or plotly
}
\description{
Plot Entlastung Einkommensverteilung
}
\examples{
\dontrun{
gen_plot_EntlastungEinkommensverteilung(SPD2025_df)}
}
