% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_plot_GrenzDurchschnittssteuersatz.R
\name{gen_plot_GrenzDurchschnittssteuersatz}
\alias{gen_plot_GrenzDurchschnittssteuersatz}
\title{Plot Grenz- und Durchschnittssteuersatz}
\usage{
gen_plot_GrenzDurchschnittssteuersatz(
  df = NULL,
  max_x = 125000,
  max_y = 60,
  fast = FALSE,
  breaks = c(0, 30000, 60000, 90000, 120000),
  interactive = FALSE
)
}
\arguments{
\item{df}{dataframe des Reformvorschlags}

\item{max_x}{Beschränkung X-Achse (ZvE)}

\item{max_y}{Beschränkung Y-Achse (Steuersatz)}

\item{breaks}{}
}
\value{
plotgen
}
\description{
Plot Grenz- und Durchschnittssteuersatz
}
\examples{
\dontrun{gen_plot_GrenzDurchschnittssteuersatz(SPD2025_df)}
}
