% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_plot_relativeEntlastung.R
\name{gen_plot_relativeEntlastung}
\alias{gen_plot_relativeEntlastung}
\title{Abbildung relative Entlastung}
\usage{
gen_plot_relativeEntlastung(
  df = NULL,
  breaks = c(10000, 30000, 1e+05, 5e+05),
  fast = FALSE,
  limit_x = 5e+05,
  limit_y = NULL
)
}
\arguments{
\item{df}{dataframe}

\item{breaks}{breaks x-Axis}

\item{limit}{limit x-Axis}

\item{limitneg_y}{lower bound y-Axis}

\item{limitpos_y}{upper bound y-Axis}
}
\value{
ggplot
}
\description{
Abbildung relative Entlastung
}
\examples{
\dontrun{
gen_plot(relativeEntlastung(SPD2025_df))}

}
