% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_waterfall_cost.R
\name{gen_plot_WaterfallCost}
\alias{gen_plot_WaterfallCost}
\title{Wasserfall plot}
\usage{
gen_plot_WaterfallCost(
  data,
  Jahre = c("2025", "2026", "2027", "2028"),
  max_y = NULL,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{max_y}{Obergrenze y-Achse}

\item{title}{Titel}

\item{subtitle}{Untertitel (e.g. Einheit)}

\item{data_Wachstumseffekte}{Dataframe Wachstumseffekte}

\item{year}{Jahr (muss enthalten sein in )}
}
\value{
ggplot
}
\description{
Wasserfall plot
}
\examples{
\dontrun{
Wasserfall_plot(Reform2025_df)}
}
