% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dz.R
\name{theme_dz}
\alias{theme_dz}
\title{Ggplot theme Dezernat Zukunft}
\usage{
theme_dz(base_size = base_size)
}
\arguments{
\item{base_size}{base size text}
}
\value{
theme
}
\description{
Ggplot theme Dezernat Zukunft
}
\examples{
\notrun{
theme_dz()}
}
