% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Grenzsteuersatz.R
\name{compute_Grenzsteuersatz}
\alias{compute_Grenzsteuersatz}
\title{Compute Grenzsteuersatz}
\usage{
compute_Grenzsteuersatz(
  zve,
  gfb = 11604,
  pg1s = 14,
  pg1e = 17005,
  pg2s = 23.97,
  pg2e = 66760,
  pp1s = 42,
  pp1e = 277826,
  pp2s = 45
)
}
\arguments{
\item{zve}{Zu versteuerndes Einkommen}

\item{gfb}{Grundfreibetrags}

\item{pg1s}{Startsatz erste Progressionszone}

\item{pg1e}{Ende erste Progressionszone}

\item{pg2s}{Startsatz zweite Progressionszone}

\item{pg2e}{Ende zweite Progressionszone}

\item{pp1s}{Satz erste Proportionalzone (Spitzensteuersatz)}

\item{pp1e}{Ende erste Proportionalzone}

\item{pp2s}{Seite zweite Proportionalzone (Reichensteuersatz)}
}
\value{
number
}
\description{
*compute_Grenzsteuersatz* berechnet den Grenzsteuersatz bei einem zu versteuernden Einkommen von *zve* basierend auf den gegebenen Eckwerten der Grenzsteuer.
}
\examples{
compute_Grenzsteuersatz(30000)
}
