% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_plot_EntlastungEinkommensverteilung.R
\name{gen_plot_EntlastungEinkommensverteilung}
\alias{gen_plot_EntlastungEinkommensverteilung}
\title{Plot Entlastung Einkommensverteilung}
\usage{
gen_plot_EntlastungEinkommensverteilung(
  df,
  breaks = c(10000, 30000, 1e+05, 5e+05, 1e+06),
  limit = 5e+05,
  width = 1,
  min_val = -4,
  max_val = 4
)
}
\arguments{
\item{df}{dataframe Reform}

\item{breaks}{breaks X-Achse}

\item{limit}{limit X-Achse}

\item{width}{breite Gruppen}

\item{min_val}{Obergrenze Gruppe}

\item{max_val}{}
}
\value{
ggplot
}
\description{
Plot Entlastung Einkommensverteilung
}
\examples{
\dontrun{
gen_plot_EntlastungEinkommensverteilung(SPD2025_df)}
}
