% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dz.R
\name{theme_dz}
\alias{theme_dz}
\title{Ggplot theme Dezernat Zukunft}
\usage{
theme_dz(base_size = 10, base_family = "Open Sans")
}
\arguments{
\item{base_size}{base size text}

\item{base_family}{base font family text}
}
\value{
theme
}
\description{
Ggplot theme Dezernat Zukunft
}
\examples{
\dontrun{
theme_dz()}
}
